+++ hexagonal/src/main/java/com/iluwatar/hexagonal/service/LotteryService.java
@@ -20,33 +20,23 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.hexagonal.domain;
+package com.iluwatar.hexagonal.service;
 
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.Optional;
+
+import com.iluwatar.hexagonal.domain.LotteryNumbers;
+import com.iluwatar.hexagonal.domain.LotteryTicket;
+import com.iluwatar.hexagonal.domain.LotteryTicketCheckResult;
+import com.iluwatar.hexagonal.domain.LotteryTicketId;
 
 /**
  * 
- * Utilities for lottery tests
+ * Interface for submitting and checking lottery tickets.
  *
  */
-public class LotteryTestUtils {
+public interface LotteryService {
 
-  /**
-   * @return lottery ticket
-   */
-  public static LotteryTicket createLotteryTicket() {
-    return createLotteryTicket("foo@bar.com", "12231-213132", "+99324554", new HashSet<>(Arrays.asList(1, 2, 3, 4)));
-  }
-  
-  /**
-   * @return lottery ticket
-   */
-  public static LotteryTicket createLotteryTicket(String email, String account, String phone,
-      Set<Integer> givenNumbers) {
-    PlayerDetails details = PlayerDetails.create(email, account, phone);
-    LotteryNumbers numbers = LotteryNumbers.create(givenNumbers);
-    return LotteryTicket.create(details, numbers);
-  }
+  Optional<LotteryTicketId> submitTicket(LotteryTicket ticket);
+
+  LotteryTicketCheckResult checkTicketForPrize(LotteryTicketId id, LotteryNumbers winningNumbers);
 }
